feat(sql): vectorized aggregation of `short` columns

+++ core/src/main/c/share/vec_agg.cpp
@@ -614,11 +614,11 @@ LONG_LONG_DISPATCHER(maxLong)
 
 extern "C" {
 
-    JNIEXPORT jlong JNICALL Java_io_questdb_std_Vect_sumShort(JNIEnv *env, jclass cl, jlong pLong, jlong count) {
+    JNIEXPORT jint JNICALL Java_io_questdb_std_Vect_sumShort(JNIEnv *env, jclass cl, jlong pLong, jlong count) {
         return sumShort_Vanilla((int16_t *) pLong, count);
     }
 
-    JNIEXPORT jlong JNICALL Java_io_questdb_std_Vect_minShort(JNIEnv *env, jclass cl, jlong pLong, jlong count) {
+    JNIEXPORT jint JNICALL Java_io_questdb_std_Vect_minShort(JNIEnv *env, jclass cl, jlong pLong, jlong count) {
         return minShort_Vanilla((int16_t *) pLong, count);
     }
